Abstract away the steps of the pipeline
