+++ time-series-forecast/src/steps/preparator.py
@@ -1,0 +1,13 @@
+from zenml.steps import step, Output
+import pandas as pd
+
+@step
+def preparator(data: pd.DataFrame) -> Output(
+    df = pd.DataFrame #np.ndarray #, X_test=np.ndarray, y_train=np.ndarray, y_test=np.ndarray
+):
+    df = data.drop(['Source_time','Lead_hours','ANM','Non_ANM', 'int64_field_0'],axis=1)
+    df = df[df['Direction'].notna()]
+    df = df[df['Total'].notna()]
+    df['Speed'] =  df['Speed'].fillna(df['Speed'].median())
+
+    return df
